+++ src/main/java/com/ludo/study/studymatchingplatform/study/repository/recruitment/position/PositionJpaRepository.java
@@ -1,10 +1,10 @@
-package com.ludo.study.studymatchingplatform.study.repository;
+package com.ludo.study.studymatchingplatform.study.repository.recruitment.position;
 
 import java.util.Set;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.ludo.study.studymatchingplatform.study.domain.Position;
+import com.ludo.study.studymatchingplatform.study.domain.recruitment.position.Position;
 
 public interface PositionJpaRepository extends JpaRepository<Position, Long> {
 	Set<Position> findByIdIn(Set<Long> positionIds);
