+++ src/main/java/com/ludo/study/studymatchingplatform/study/repository/recruitment/stack/StackCategoryJpaRepository.java
@@ -1,0 +1,9 @@
+package com.ludo.study.studymatchingplatform.study.repository.recruitment.stack;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.ludo.study.studymatchingplatform.study.domain.recruitment.stack.StackCategory;
+
+public interface StackCategoryJpaRepository extends JpaRepository<StackCategory, Long> {
+
+}
