+++ src/main/java/com/ludo/study/studymatchingplatform/study/repository/study/StudyJpaRepository.java
@@ -1,8 +1,8 @@
-package com.ludo.study.studymatchingplatform.study.repository;
+package com.ludo.study.studymatchingplatform.study.repository.study;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.ludo.study.studymatchingplatform.study.domain.Study;
+import com.ludo.study.studymatchingplatform.study.domain.study.Study;
 
 public interface StudyJpaRepository extends JpaRepository<Study, Long> {
