+++ src/main/java/com/ludo/study/studymatchingplatform/study/repository/study/StudyRepositoryImpl.java
@@ -1,15 +1,15 @@
-package com.ludo.study.studymatchingplatform.study.repository;
+package com.ludo.study.studymatchingplatform.study.repository.study;
 
-import static com.ludo.study.studymatchingplatform.study.domain.QParticipant.*;
-import static com.ludo.study.studymatchingplatform.study.domain.QStudy.*;
 import static com.ludo.study.studymatchingplatform.study.domain.recruitment.QRecruitment.*;
-import static com.ludo.study.studymatchingplatform.user.domain.QUser.*;
+import static com.ludo.study.studymatchingplatform.study.domain.study.QParticipant.*;
+import static com.ludo.study.studymatchingplatform.study.domain.study.QStudy.*;
+import static com.ludo.study.studymatchingplatform.user.domain.user.QUser.*;
 
 import java.util.Optional;
 
 import org.springframework.stereotype.Repository;
 
-import com.ludo.study.studymatchingplatform.study.domain.Study;
+import com.ludo.study.studymatchingplatform.study.domain.study.Study;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 
 import lombok.RequiredArgsConstructor;
