+++ src/main/java/com/ludo/study/studymatchingplatform/study/repository/study/category/CategoryJpaRepository.java
@@ -1,0 +1,9 @@
+package com.ludo.study.studymatchingplatform.study.repository.study.category;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.ludo.study.studymatchingplatform.study.domain.study.Category;
+
+public interface CategoryJpaRepository extends JpaRepository<Category, Long> {
+
+}
