+++ src/main/java/com/ludo/study/studymatchingplatform/study/repository/study/participant/ParticipantJpaRepository.java
@@ -1,8 +1,8 @@
-package com.ludo.study.studymatchingplatform.study.repository;
+package com.ludo.study.studymatchingplatform.study.repository.study.participant;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.ludo.study.studymatchingplatform.study.domain.Participant;
+import com.ludo.study.studymatchingplatform.study.domain.study.Participant;
 import com.ludo.study.studymatchingplatform.study.domain.id.ParticipantId;
 
 public interface ParticipantJpaRepository extends JpaRepository<Participant, ParticipantId> {
