+++ src/main/java/com/ludo/study/studymatchingplatform/user/repository/user/UserJpaRepository.java
@@ -1,8 +1,8 @@
-package com.ludo.study.studymatchingplatform.user.repository;
+package com.ludo.study.studymatchingplatform.user.repository.user;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.ludo.study.studymatchingplatform.user.domain.User;
+import com.ludo.study.studymatchingplatform.user.domain.user.User;
 
 public interface UserJpaRepository extends JpaRepository<User, Long> {
